+++ src/DfReader.cpp
@@ -128,8 +128,15 @@ class DfReader {
   std::vector<VarType> var_types_;
   std::vector<std::string> notes_;
 
+  // If empty, assume all will be read in
+  std::set<std::string> colsOnly_;
+
 public:
   DfReader(FileType type, bool user_na = false): type_(type), nrows_(0), ncols_(0), user_na_(user_na) {
+  }
+
+  void restrictCols(const std::set<std::string>& cols) {
+    colsOnly_ = cols;
   }
 
   void setInfo(int obs_count, int var_count) {
@@ -141,7 +148,7 @@ public:
       nrowsAlloc_ = nrows_ = obs_count;
     }
 
-    ncols_ = var_count;
+    ncols_ = colsOnly_.empty() ? var_count : colsOnly_.size();
 
     output_ = List(ncols_);
     names_ = CharacterVector(ncols_);
@@ -161,25 +168,33 @@ public:
     }
   }
 
-  void createVariable(int index, readstat_variable_t *variable, const char *val_labels) {
+  int createVariable(int index, readstat_variable_t *variable, const char *val_labels) {
 
-    names_[index] = Rf_mkCharCE(readstat_variable_get_name(variable), CE_UTF8);
+    const char* name = readstat_variable_get_name(variable);
+
+    if (!colsOnly_.empty() && colsOnly_.count(name) == 0) {
+      return READSTAT_HANDLER_SKIP_VARIABLE;
+    }
+
+    int var_index = readstat_variable_get_index_after_skipping(variable);
+
+    names_[var_index] = Rf_mkCharCE(name, CE_UTF8);
 
     switch(readstat_variable_get_type(variable)) {
     case READSTAT_TYPE_STRING_REF:
     case READSTAT_TYPE_STRING:
-      output_[index] = CharacterVector(nrowsAlloc_);
+      output_[var_index] = CharacterVector(nrowsAlloc_);
       break;
     case READSTAT_TYPE_INT8:
     case READSTAT_TYPE_INT16:
     case READSTAT_TYPE_INT32:
     case READSTAT_TYPE_FLOAT:
     case READSTAT_TYPE_DOUBLE:
-      output_[index] = NumericVector(nrowsAlloc_);
+      output_[var_index] = NumericVector(nrowsAlloc_);
       break;
     }
 
-    RObject col = output_[index];
+    RObject col = output_[var_index];
 
     const char* var_label = readstat_variable_get_label(variable);
     if (var_label != NULL && strcmp(var_label, "") != 0) {
@@ -187,13 +202,13 @@ public:
     }
 
     if (val_labels != NULL)
-      val_labels_[index] = val_labels;
+      val_labels_[var_index] = val_labels;
 
     const char* var_format = readstat_variable_get_format(variable);
 
     VarType var_type = numType(type_, var_format);
     // Rcout << var_name << ": " << var_format << " [" << var_type << "]\n";
-    var_types_[index] = var_type;
+    var_types_[var_index] = var_type;
     switch(var_type) {
     case HAVEN_DATE:
       col.attr("class") = "Date";
@@ -279,10 +294,13 @@ public:
     if (type_ == HAVEN_SPSS && display_width != 8) {
       col.attr("display_width") = Rf_ScalarInteger(display_width);
     }
+
+    return READSTAT_HANDLER_OK;
   }
 
   void setValue(int obs_index, readstat_variable_t *variable, readstat_value_t value) {
-    int var_index = variable->index;
+    int var_index = readstat_variable_get_index_after_skipping(variable);
+
     VarType var_type = var_types_[var_index];
 
     if (obs_index >= nrowsAlloc_)
@@ -402,8 +420,7 @@ int dfreader_note(int note_index, const char *note, void *ctx) {
 
 int dfreader_variable(int index, readstat_variable_t *variable,
                       const char *val_labels, void *ctx) {
-  ((DfReader*) ctx)->createVariable(index, variable, val_labels);
-  return 0;
+  return ((DfReader*) ctx)->createVariable(index, variable, val_labels);
 }
 int dfreader_value(int obs_index, readstat_variable_t *variable,
                    readstat_value_t value, void *ctx) {
@@ -618,8 +635,14 @@ List df_parse_xpt(Rcpp::List spec, std::string encoding = "") {
 
 template<typename InputClass>
 List df_parse_sas(Rcpp::List spec_b7dat, Rcpp::List spec_b7cat,
-                  std::string encoding) {
+                  std::string encoding, std::vector<std::string> cols_only) {
   DfReader builder(HAVEN_SAS);
+
+  if (!cols_only.empty()) {
+    std::set<std::string> cols_set(cols_only.begin(), cols_only.end());
+    builder.restrictCols(cols_set);
+  }
+
   InputClass builder_input_dat(spec_b7dat);
 
   readstat_parser_t* parser = haven_init_parser(encoding);
@@ -652,13 +675,13 @@ List df_parse_sas(Rcpp::List spec_b7dat, Rcpp::List spec_b7cat,
 
 // [[Rcpp::export]]
 List df_parse_sas_file(Rcpp::List spec_b7dat, Rcpp::List spec_b7cat,
-                       std::string encoding) {
-  return df_parse_sas<DfReaderInputFile>(spec_b7dat, spec_b7cat, encoding);
+                       std::string encoding, std::vector<std::string> cols_only) {
+  return df_parse_sas<DfReaderInputFile>(spec_b7dat, spec_b7cat, encoding, cols_only);
 }
 // [[Rcpp::export]]
 List df_parse_sas_raw(Rcpp::List spec_b7dat, Rcpp::List spec_b7cat,
-                      std::string encoding) {
-  return df_parse_sas<DfReaderInputRaw>(spec_b7dat, spec_b7cat, encoding);
+                      std::string encoding, std::vector<std::string> cols_only) {
+  return df_parse_sas<DfReaderInputRaw>(spec_b7dat, spec_b7cat, encoding, cols_only);
 }
 
 // [[Rcpp::export]]